+++ new.js
@@ -1,0 +1,41 @@
+// first thing they did was create a duplicate of the SOT by doing newDate = data.slice()
+
+//keep track stuff is the state ...
+
+// how many times voted
+let totalVotes = 0;
+
+// a vote is: 
+
+// keep track of votes for a given item
+let voteDetails = []
+  /* 
+    need ids ... like shopping cart, but instead of quantity, it's votes;
+    */
+
+// 
+const displayThreeObjects = () => {
+    //need somewhere to store products
+    //get the three random products:
+        const randomNum = math.random( ) * productsData.length
+        const ran
+    //make sure they are unique
+    //need to render on screen as radio buttons (radios with same name and different values)
+    
+    //EVENT LISTENER -- on submit = () => 
+    //add event listener to select on the the three products
+    //upon select update total votes, and the vote count for each product
+        // this is complicated...so they say
+          // incrememnet votes for item in the array
+            //but first need to check if the coffee exists in votes array, if not, push the coffee into the array
+}
+
+// reset the whole app when finished
+    // reset the votes array and total votes to there initial states
+        // or, set the votes array and total votes to their intial statial
+
+/* const intializeState = () => {
+  totalVotes = 0;
+  productVoteDetails = []
+}
+
