+++ style.css
@@ -1,12 +1,19 @@
 #survey {
-  display: flex;
-  flex-direction: row;
-  justify-content: space-evenly;
+
 }
 
 #submit-button-div {
+  
+}
+
+#img-div {
   display: flex;
-  align-items: center;
-  justify-content: center;
-  margin-top: 35px;
+  flex-direction: row;
+  align-content: space-around;
+}
+
+.photos {
+  display: flex;
+  flex-direction: row;
+  align-content: space-around;
 }
